// src/modules/risk-profiler/interfaces/response.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ResponseSchema {
  auto: "economic" | "regular" | "responsible" | "ineligible";
  disability: "economic" | "regular" | "responsible" | "ineligible";
  home: "economic" | "regular" | "responsible" | "ineligible";
  life: "economic" | "regular" | "responsible" | "ineligible";
}